# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Nimpass REST API
  version: 1.0.0
  license: { name: MIT, url: 'https://spdx.org/licenses/MIT.html' }
  description: |
    All money values are integer Luna. Login and payout proofs must be produced
    by Nimiq Pay's sign(message), using the exact server-provided challenge message.
    Payout completion requires signatures by both the new payout wallet and the
    authenticated owner wallet over that same challenge message.
    Unsafe methods require an exact configured Origin (or Referer fallback).
    Shared middleware errors use the Error schema: ORIGIN_FORBIDDEN (403),
    AUTH_REQUIRED (401), CSRF_INVALID (403), CONTENT_TYPE (415),
    VALIDATION_ERROR (400), RATE_LIMITED (429) and INTERNAL_ERROR (500).
    Session lookup infrastructure failures return INTERNAL_ERROR (500), preserving
    the cookie; AUTH_REQUIRED (401) means a missing, invalid, expired or revoked session.
    Authenticated operations require a session; every unsafe authenticated
    operation requires CSRF. JSON bodies reject unknown fields and trailing data.
    Text length bounds are UTF-8 byte limits after documented normalization.
    Authenticated unsafe methods additionally require X-CSRF-Token, obtained
    from login or GET /auth/session. Cookies are HttpOnly, SameSite=Lax; HTTPS
    deployments use Secure __Host-nimpass_session, local HTTP development uses
    nimpass_session. Requests must send credentials: include.
...
    unchanged.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /auth/challenges:
    post:
      summary: Create wallet login challenge
      operationId: createLoginChallenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChallengeInput' }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Challenge created, content: { application/json: { schema: { $ref: '#/components/schemas/Challenge' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /auth/sessions:
    post:
      summary: Verify wallet signature and create session
      operationId: completeLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Proof' }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Authenticated; sets session cookie, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /auth/session:
    get:
      summary: Get current session
      operationId: currentSession
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Current identity and CSRF token, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Error' }
    delete:
      summary: Revoke current session
      operationId: logout
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        default: { $ref: '#/components/responses/Error' }
        '204': { description: Session revoked; clears cookie }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /providers/{providerID}/payout-challenges:
    parameters: [ { $ref: '#/components/parameters/ProviderID' } ]
    post:
      summary: Create payout wallet challenge
      operationId: createPayoutChallenge
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChallengeInput' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '201': { description: Purpose VERIFY_PROVIDER_WALLET challenge, content: { application/json: { schema: { $ref: '#/components/schemas/Challenge' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /providers/{providerID}/payout-verifications:
    parameters: [ { $ref: '#/components/parameters/ProviderID' } ]
    post:
      summary: Verify payout wallet and owner step-up
      operationId: verifyPayout
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PayoutProof' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Verified provider payout wallet, content: { application/json: { schema: { $ref: '#/components/schemas/Provider' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-nimpass_session
      description: HTTPS production cookie; local HTTP development uses nimpass_session.
  parameters:
    Csrf: { in: header, name: X-CSRF-Token, required: true, schema: { type: string, pattern: '^[0-9a-f]{64}$' } }
    ProviderID: { in: path, name: providerID, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Normalized error; codes include AUTH_REQUIRED, INVALID_SIGNATURE, CHALLENGE_EXPIRED, CHALLENGE_CONSUMED, FORBIDDEN, VALIDATION_ERROR, NOT_FOUND, CONFLICT, PAYMENT_CONFLICT, PASS_UNAVAILABLE, PASS_ALREADY_OWNED, PURCHASE_IN_SETTLEMENT, PASS_PURCHASE_CUTOFF, SELF_PURCHASE_NOT_ALLOWED, INTENT_EXPIRED, PASS_NOT_FOUND, PASS_NOT_OWNED, PASS_EXPIRED, PASS_COMPLETED, REDEMPTION_CHALLENGE_EXPIRED, REDEMPTION_ALREADY_CONSUMED, REDEMPTION_NOT_AUTHORIZED, INVALID_REDEMPTION_SIGNATURE, STALE_REDEMPTION_CHALLENGE, INVALID_REDEMPTION_TOKEN, RATE_LIMITED, CSRF_INVALID, ORIGIN_FORBIDDEN.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    ChallengeInput:
      type: object
      additionalProperties: false
      required: [wallet]
      properties: { wallet: { type: string, description: Nimiq NQ address; spaces accepted } }
    Challenge:
      type: object
      required: [id, purpose, wallet, message, expiresAt]
      properties:
        id: { type: string, format: uuid }
        purpose: { type: string, enum: [AUTH_LOGIN, VERIFY_PROVIDER_WALLET] }
        wallet: { type: string }
        message: { type: string, description: Sign this exact UTF-8 string; do not reconstruct it. }
        expiresAt: { type: string, format: date-time }
...
    SigningScheme:
      type: string
      enum: [raw, hub]
      description: >-
        Which documented preprocessing the wallet applied before signing.
        Omit it to verify under the deployment's configured NIMIQ_SIGNING_SCHEME.
        Send "hub" for a signature produced by @nimiq/hub-api signMessage(),
        whose envelope is specified as sha256("\x16Nimiq Signed Message:\n" +
        length + message). The Nimiq Pay Mini App provider does not document its
        preprocessing, so that transport sends nothing. Exactly one scheme is
        applied per proof; the server never retries under another.
    Proof:
      type: object
      additionalProperties: false
      required: [challengeId, wallet, publicKey, signature]
      properties:
        challengeId: { type: string, format: uuid }
        wallet: { type: string, description: Nimiq NQ address; spaces accepted }
        publicKey: { type: string, pattern: '^[0-9a-fA-F]{64}$' }
        signature: { type: string, pattern: '^[0-9a-fA-F]{128}$' }
        signingScheme: { $ref: '#/components/schemas/SigningScheme' }
    PayoutProof:
      type: object
      additionalProperties: false
      required: [challengeId, wallet, publicKey, signature, ownerPublicKey, ownerSignature]
      properties:
        challengeId: { type: string, format: uuid }
        wallet: { type: string, description: Nimiq NQ address; spaces accepted }
        publicKey: { type: string, pattern: '^[0-9a-fA-F]{64}$' }
        signature: { type: string, pattern: '^[0-9a-fA-F]{128}$' }
        ownerPublicKey: { type: string, pattern: '^[0-9a-fA-F]{64}$' }
        ownerSignature: { type: string, pattern: '^[0-9a-fA-F]{128}$' }
        signingScheme: { $ref: '#/components/schemas/SigningScheme' }
    Identity:
      type: object
      required: [id, wallet, createdAt]
      properties:
        id: { type: string, format: uuid }
        wallet: { type: string }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      required: [identity, expiresAt, csrfToken]
      properties:
        identity: { $ref: '#/components/schemas/Identity' }
        expiresAt: { type: string, format: date-time }
        csrfToken: { type: string, pattern: '^[0-9a-f]{64}$' }
    ProviderSlug:
      type: string
      minLength: 3
      maxLength: 80
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      description: Trimmed and lowercased ASCII slug. UUIDs and reserved route names are rejected. Assigned once at creation; changing name never changes the slug. If omitted, a readable name prefix plus UUID suffix is generated. Duplicate custom slugs return CONFLICT.
...
    Provider:
      type: object
      required: [id, name, slug, headline, bio, avatarUrl, avatarVariant, location, payoutWallet, payoutVerifiedAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { $ref: '#/components/schemas/ProviderSlug' }
        headline: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        avatarVariant: { type: integer, minimum: 0, maximum: 255 }
        location: { type: string }
        payoutWallet: { type: string, nullable: true }
        payoutVerifiedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }